Simplify makefile
There is no need to specify the compiler invocations, there are already
built-in rules for that. All you really need to do is set the standard
variables, and let make take care of the rest. And add an "all" default
target, which allows to call multiple target like "make all clean"
without having to use any specific executable-/file-names.

Also remove the "copy" of the makefile from the readme, and only
reference the one in the source dir. And reduce the required package
list to the actual required minimum.