Use correct fps mpv option
In mpv 0.37.0 the `fps` option/property was renamed to
`container-fps-override` [0][1], and a compatibility alias was added so
both the new and old names worked. In the recently released mpv 0.39.0
that compatibility alias was removed [2][3], using the old name does not
work any more. Since VideoCut is currently using the old name, it does
currently not work with the latest mpv release.

mpv 0.37.0 was released in November 2023, so it is less than a year old.
Most distributions (like Debian stable) will ship a mpv version that
requires the old name, while other (like Arch) already ship a version that
requires the new name, which makes it necessary to support both names.

The solution is a simple version check to decide which name to use. I
choose 0.39.0 as the switch-over point, as to not change the current
behaviour.

[0] https://github.com/mpv-player/mpv/blob/release/0.37/DOCS/interface-changes.rst?plain=1#L103
[1] https://github.com/mpv-player/mpv/commit/7aed492ccc48fd32084d7dfe65a50b5b90d8411a
[2] https://github.com/mpv-player/mpv/blob/release/0.39/DOCS/interface-changes.rst?plain=1#L41
[3] https://github.com/mpv-player/mpv/commit/6e3d90d72a995bfb8b91d79538dd354c6221d527